// src/app/models/employee.ts
export interface Employee {
  employeeId: number;
  firstName: string;
  lastName: string;
  departmentId: number;
  departmentName: string;
  address: string;
  idProofTypeId: number;
  idProofType: string;
  idProof: string;
  salary: number;
  salaryCategoryId: number;
  salaryCategoryName: string;
  dateOfJoining: string;
  contactNo: number;
  allowedLeaveDays: number;
}